mo args mo problems

+++ scripts/renderByRegexp.py
@@ -8,6 +8,7 @@ renderPreset = "MP4 1080p"
 
 parser = argparse.ArgumentParser(description="Add timelines to Resolve's render queue by specifying a suffix.")
 parser.add_argument("-d", "--dest", help="Destination directory for the renders")
+parser.add_argument("-k", "--keep", help="Keep the queue instead of deleting it")
 parser.add_argument("regexp")
 args = parser.parse_args()
 
@@ -15,7 +16,8 @@ args = parser.parse_args()
 if not args.dest:
     args.dest = os.getcwd()
 
-resolve.DeleteAllRenderJobs()
+if not args.keep:
+    resolve.DeleteAllRenderJobs()
 
 for tl in resolve.GetTimelinesByRegexp(args.regexp):
     resolve.AddTimelineToRender(tl, renderPreset, args.dest)
